moved truck model to shared lib
