+++ AKS_IoTHub_Reader/telemetryReader/telemetryReader.csproj
@@ -24,4 +24,8 @@
     </None>
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\telemetryModel\telemetryModel.csproj" />
+  </ItemGroup>
+
 </Project>
