+++ AKS_IoTHub_Reader/telemetryReader/telemetryreader.yaml
@@ -16,7 +16,7 @@ spec:
     spec:
       containers:
       - name: telemetryreader
-        image: aksiottest.azurecr.io/telemetryreader:1.0.4
+        image: aksiottest.azurecr.io/telemetryreader:1.0.5
         stdin: true
         ports:
         - containerPort: 443
