+++ AKS_IoTHub_Reader/telemetrySimulator/telemetrySimulator.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\telemetryModel\telemetryModel.csproj" />
+  </ItemGroup>
+
 </Project>
